Move server-side worker ID tests into test_worker_id.py
